+++ services/master/controllers/admin.go
@@ -10,7 +10,3 @@ func HelloAdmin(c *fiber.Ctx) error {
 	msg := fmt.Sprintf("Hello, admin %s 👋!", c.Params("name"))
 	return c.SendString(msg)
 }
-
-func Test() {
-
-}
